==> 1-scrape-elevation.ipynb <==
### Import the required packages


import mechanicalsoup as soup
import pandas as pd

### Scrape the elevation from the url


# Create a url variable

url = "https://en.m.wikipedia.org/wiki/List_of_countries_by_average_elevation"

# create a browser object

browser = soup.StatefulBrowser()

# open the url

browser.open(url)

# get the page

page = browser.get_current_page()

# get the table tag with the class 'wikitable sortable'

table = page.find("table", class_="wikitable sortable")

# get the rows using the tr tag

rows = table.find_all("tr")

### Inspect the Scraped Data


# print the rows

for i in range(len(rows)):
    print(rows[i].text)

# We see that the first row is the header in str format separated by "\n"

print(rows[0].text)

# And the second row is the Country name followed by the average elevation("m" and "feet") again separated by "\n"

print(rows[1].text)

### Load the scraped data into a dataframe


# Create a dataframe and append the rows in the table

df = pd.DataFrame(columns=["Country-elevations"])

# Create a loop and append the scraped rows in the table

for i in range(1, len(rows)):
    new_df = pd.DataFrame([[rows[i].text]], columns=["Country-elevations"])

    df = pd.concat([df, new_df], ignore_index=True)

df.head(5)

### Data Cleaning


#### Clean the Elevation Column

# Get rid of the trailing characters "(6181 ft)[2]" in the Country-elevation column

# 1. We can use the split() method to split the string and get the first element as the Country name

df["Country"] = df["Country-elevations"].apply(lambda x: x.split("\n")[0])

# 2. Similarly, we can use the split() method to split the string and get the second element as the average elevation

df["Average-elevation"] = df["Country-elevations"].apply(lambda x: x.split("\n")[1])

df.head(1)

# 3. We can drop the trailing characters: " m (6181 ft)[2]" in the Country-elevation column

df["Numeric-Elevation"] = df["Average-elevation"].str.split("m").str[0].str.rstrip()

df.head(1)

# 4. finally, we can get rid of the "," in the numeric-elevation column and convert the column to float
df["Numeric-Elevation"] = df["Numeric-Elevation"].str.replace(",", "").astype(float)

df.head(1)

# Drop the 'Country-elevations' and 'Average-elevation' columns

df.drop(["Country-elevations", "Average-elevation"], axis=1, inplace=True)

df.head(1)

#### Clean the Country Column

# Inspect the Country column strings

df["Country"][0]


# Clean the leading white spaces in the Country column

df["Country"] = df["Country"].str.strip()

# Inspect the Country column strings

df["Country"][0]

### Save the dataframe as a CSV file

# Save the dataframe as a csv file

df.to_csv("clean-elevation-data.csv", index=False)
