==> 2-csv-get-coordinates.ipynb <==
### Load and Inspect "Countries-Coordinates" CSV format Data


# Read semi-colon separated file with geolocation data
import pandas as pd

# url of csv file
url = "https://public.opendatasoft.com/api/explore/v2.1/catalog/datasets/ne_10m_admin_0_countries/exports/csv?lang=en&timezone=Europe%2FBerlin&use_labels=true&delimiter=%3B"

# Use pandas to read the csv file from the url
df_geo = pd.read_csv(url, sep=";")

# df_geo = pd.read_csv("raw-countries-coordinates.csv", sep=";") # read from downloaded file if you have no internet connection

# print the shape of the Geo data
print(df_geo.shape)

# Show first 5 rows
df_geo.head()

### Data Cleaning


# Rename columns to match the other dataframes
df_geo.rename(
    columns={
        "SOVEREIGNT": "Country",
        "Geo Point": "coordinates",
        "SOV_A3": "Country-code",
        "CONTINENT": "continent",
    },
    inplace=True,
)

# Show first rows
df_geo.head(3)

# Drop all columns except Country, coordinates and continent

# Create a list of columns
columns_list = ["Country", "coordinates", "continent"]

# Drop all columns except the ones in the list
df_geo = df_geo[columns_list]

# Show first 5 rows
df_geo.head()

# Filter only European countries
df_geo_eur = df_geo[df_geo["continent"] == "Europe"]

# checking for duplicates in column Country
df_geo_eur[df_geo_eur.duplicated(subset=["Country"])]

# Drop duplicates in column Country
df_geo_eur.drop_duplicates(subset=["Country"], inplace=True)

# checking for duplicates in column Country
df_geo_eur[df_geo_eur.duplicated(subset=["Country"])]

df_geo_eur["Country"].__len__()

# find country Cyprus in the dataframe countries

df_geo_eur[df_geo_eur["Country"] == "Cyprus"]

# Add missing country Cyprus to the dataframe using concat

# Create a DataFrame for the target Country
target_country_df = pd.DataFrame(
    {"Country": ["Cyprus"], "coordinates": [
        "34.977, 33.106"], "continent": ["Europe"]}
)

# Concatenate the DataFrames
df_geo_eur = pd.concat([df_geo_eur, target_country_df], ignore_index=True)

# Check if Cyprus is in the dataframe
df_geo_eur[df_geo_eur["Country"] == "Cyprus"]

# Save dataframe to csv file
df_geo_eur.to_csv("clean-coordinates-data.csv", index=False)
