==> 3-merge-coordinates-elevation.ipynb <==
### Load the Coordinates and Elevation CSV files

import pandas as pd

# Load the coordinate csv file

df_coordinates = pd.read_csv("clean-coordinates-data.csv")

df_coordinates.head(2)


# Load the elevation csv file

df_elevation = pd.read_csv("clean-elevation-data.csv")

df_elevation.head(2)

### Dataset Validation

# Check if all Country names columns in both dataframes are the same

# Check if elements of coordinates['Country'] are in df_elevation['Country']
df_coordinates["in_elevation"] = df_coordinates["Country"].isin(
    df_elevation["Country"])

# Inspect the result
df_coordinates[df_coordinates["in_elevation"] == False]


**In Coordinates dataset, Rename the following countries**:
1.  Czechia as Czech Republic.
2.  Republic of Serbia as Serbia.
3.  Macedonia as North Macedonia.
4.  Vatican as Vatican City.

**In Elevation dataset, For Microstates, we will impute the average elevation of the Country they are within:**
- San Marino is a microstate within Italy. It is not present in the elevation dataset. 
- Monaco is a microstate within France.
- Vatican City is a microstate within Italy.
- Liechtenstein is a microstate within Switzerland.
  
**Malta is an island nation in the Mediterranean Sea with Average elevation of 127 meters. Add this value to the elevation dataset**


#### Renaming Czechia, Republic of Serbia, Macedonia, Vatican in Coordinates dataset

# Create a function to rename countries in a DataFrame


def rename_countries(df, rename_dict):
    """
    Renames countries in the DataFrame based on a provided dictionary.

    Parameters:
    df (pd.DataFrame): The DataFrame with Country data.
    rename_dict (dict): A dictionary where keys are current Country names and
                        values are new Country names.

    Returns:
    pd.DataFrame: The DataFrame with updated Country names.
    """
    for current_name, new_name in rename_dict.items():
        df.loc[df["Country"] == current_name, "Country"] = new_name

    return df


rename_dict = {
    "Czechia": "Czech Republic",
    "Republic of Serbia": "Serbia",
    "Macedonia": "North Macedonia",
    "Vatican": "Vatican City",
}

df_coordinates = rename_countries(df_coordinates, rename_dict)

# Check results
print(df_coordinates[df_coordinates["Country"] == "Czech Republic"])

print(df_coordinates[df_coordinates["Country"] == "Serbia"])

print(df_coordinates[df_coordinates["Country"] == "North Macedonia"])

print(df_coordinates[df_coordinates["Country"] == "Vatican City"])

#### Microstates Imputation using their respective countries' average elevation

# Create a function to impute elevation data for a given Country


def add_country_with_another_elevation(df, source_country, target_country):
    """
    Adds a new row to the DataFrame for the target Country with the elevation of the source Country.

    Parameters:
    df (pd.DataFrame): The DataFrame to modify.
    source_country (str): The name of the Country whose elevation is to be copied.
    target_country (str): The name of the new Country to be added.

    Returns:
    pd.DataFrame: The modified DataFrame.
    """
    # Check if source Country is in the DataFrame
    if source_country in df["Country"].values:
        # Find source Country's elevation
        source_elevation = df.loc[
            df["Country"] == source_country, "Numeric-Elevation"
        ].iloc[0]

        # Create a DataFrame for the target Country
        target_country_df = pd.DataFrame(
            {"Country": [target_country],
                "Numeric-Elevation": [source_elevation]}
        )

        # Concatenate the existing DataFrame with the target Country DataFrame
        return pd.concat([df, target_country_df], ignore_index=True)
    else:
        print(f"{source_country} not found in the DataFrame.")
        return df

# Impute elevation data for San Marino, Monaco, Vatican City, and Liechtenstein

df_elevation = add_country_with_another_elevation(
    df_elevation, "Italy", "San Marino")

df_elevation = add_country_with_another_elevation(
    df_elevation, "France", "Monaco")

df_elevation = add_country_with_another_elevation(
    df_elevation, "Italy", "Vatican City")

df_elevation = add_country_with_another_elevation(
    df_elevation, "Switzerland", "Liechtenstein"
)

# Check results
print(df_elevation[df_elevation["Country"] == "San Marino"])

print(df_elevation[df_elevation["Country"] == "Monaco"])

print(df_elevation[df_elevation["Country"] == "Vatican City"])

print(df_elevation[df_elevation["Country"] == "Liechtenstein"])

### Add the Country "Malta" with average elevation of 127 meters to the elevation dataset.

# Create a function to add a new country with a given elevation


def add_country_with_elevation(df, country, elevation):
    """
    Adds a new row to the DataFrame for a specified Country with a given elevation.

    Parameters:
    df (pd.DataFrame): The DataFrame to modify.
    country (str): The name of the country to be added.
    elevation (float): The elevation value for the country.

    Returns:
    pd.DataFrame: The modified DataFrame.
    """
    # Create a DataFrame for the new country
    new_country_df = pd.DataFrame(
        {"Country": [country], "Numeric-Elevation": [elevation]}
    )

    # Concatenate the existing DataFrame with the new country DataFrame
    return pd.concat([df, new_country_df], ignore_index=True)


# Add Malta with an elevation of 127
df_elevation = add_country_with_elevation(df_elevation, "Malta", 127.0)

df_elevation[df_elevation["Country"] == "Malta"]

# Merge the two dataframes on the 'Country' column

merged = pd.merge(df_coordinates, df_elevation, how="inner", on="Country")

# Drop the 'in_elevation' column
merged.drop("in_elevation", axis=1, inplace=True)

# Drop the 'continent' column
merged.drop("continent", axis=1, inplace=True)

# Rename the 'Numeric-Elevation' column
merged.rename(columns={"Numeric-Elevation": "elevation"}, inplace=True)

# Reorder the columns as 'Country', 'coordinates', 'elevation', 'continent'
merged = merged[["Country", "coordinates", "elevation"]]

merged

# Save the merged dataframe to a csv file
merged.to_csv("4-final-geolocation-data.csv", index=False)
