==> 6-final-merge.ipynb <==
### Load and Inspect the Datasets

# Load AirQuality, Weather-GDP, and Geolocation data from the CSV files

import pandas as pd

df_airquality = pd.read_csv("AirQuality_RenewableEnergy_data.csv")

df_weather = pd.read_csv("weather_gdp_data.csv")

df_geo = pd.read_csv("4-final-geolocation-data.csv")

print(df_airquality.shape)

df_airquality.head(2)

print(df_weather.shape)
df_weather.head(2)

print(df_geo.shape)

df_geo.head(2)

### Merge the Datasets

#### Merge weather and geolocation datasets

df_geo_weather = pd.merge(df_geo, df_weather, on="Country", how="left")

df_geo_weather.shape

df_geo_weather.head(2)

#### Final Merge df_airquality and df_geo_weather

df_airquality_geo_weather = pd.merge(
    df_airquality, df_geo_weather, on="Country", how="left"
)

print(df_airquality_geo_weather.shape)

print(df_airquality_geo_weather.head(2))

# Display rows with missing values
df_airquality_geo_weather[df_airquality_geo_weather.isnull().any(axis=1)]

# Drop rows with missing values
df_airquality_geo_weather.dropna(inplace=True)

print(df_airquality_geo_weather.shape)

# Save the merged data to a CSV file
df_airquality_geo_weather.to_csv(
    "Final_AirQuality_Geolocation_weather.csv", index=False
)
