==> 7-mariadb_Final_studentC.ipynb <==
# Module Imports
import mariadb
import sys
import pandas as pd


# Connect to MariaDB Platform
try:
    conn = mariadb.connect(
        user="cip_user",
        password="cip_pw",
        host="127.0.0.1",                    # 10.177.124.185,  localhost , 127.0.0.1
        # host="10.177.124.35",                    # 10.177.124.35,  localhost , 127.0.0.1    FS23
        port=3306,
        database="AIRQUALITY"

    )
except mariadb.Error as e:
    print(f"Error connecting to MariaDB Platform: {e}")
    sys.exit(1)

# Get Cursor
cur = conn.cursor()

cur

cur.execute('''
CREATE TABLE IF NOT EXISTS tbl_Airquality (
    Year INT,
    Country VARCHAR(255),
    Population INT,
    Populated_Area_km2 FLOAT(10,2),
    Air_Pollution_Average_ug_m3_NO2 FLOAT(5,2),
    Air_Pollution_Average_ug_m3_O3 FLOAT(10,2),
    Air_Pollution_Average_ug_m3_PM10 FLOAT(5,2),
    Air_Pollution_Average_ug_m3_PM2_5 FLOAT(5,2),
    Annual_AQI_NO2 FLOAT(5,2),
    Annual_AQI_O3 FLOAT(5,2),
    Annual_AQI_PM10 FLOAT(5,2),
    Annual_AQI_PM2_5 FLOAT(5,2),
    Premature_Deaths_NO2 FLOAT(10,2),
    Premature_Deaths_O3 FLOAT(10,2),
    Premature_Deaths_PM10 FLOAT(10,2),
    Premature_Deaths_PM2_5 FLOAT(10,2),
    Years_Of_Life_Lost_NO2 FLOAT(10,2),
    Years_Of_Life_Lost_O3 FLOAT(10,2),
    Years_Of_Life_Lost_PM10 FLOAT(10,2),
    Years_Of_Life_Lost_PM2_5 FLOAT(10,2),
    RenewableData INT,
    Average_Renewable_Data FLOAT(5,2),
    coordinates VARCHAR(255),
    elevation FLOAT,
    Country_Code VARCHAR(10),
    New_Rank INT,
    Mean_GDP FLOAT,
    Temperature FLOAT,
    Wind_Speed FLOAT
)
''')

### Fetch data from the Final CSV file


# Load the CSV file
csv_file_path = "Final_AirQuality_Geolocation_weather.csv"
df = pd.read_csv(csv_file_path)

### Insert data into the 'tbl_Airquality' table


# Create a sql query for insertion

sql_insert_query = '''
    INSERT INTO tbl_Airquality (
    Year, 
    Country, 
    Population, 
    Populated_Area_km2, 
    Air_Pollution_Average_ug_m3_NO2, 
    Air_Pollution_Average_ug_m3_O3, 
    Air_Pollution_Average_ug_m3_PM10, 
    Air_Pollution_Average_ug_m3_PM2_5, 
    Annual_AQI_NO2, 
    Annual_AQI_O3, 
    Annual_AQI_PM10, 
    Annual_AQI_PM2_5, 
    Premature_Deaths_NO2, 
    Premature_Deaths_O3, 
    Premature_Deaths_PM10, 
    Premature_Deaths_PM2_5, 
    Years_Of_Life_Lost_NO2, 
    Years_Of_Life_Lost_O3, 
    Years_Of_Life_Lost_PM10, 
    Years_Of_Life_Lost_PM2_5, 
    RenewableData, 
    Average_Renewable_Data,
    coordinates,
    elevation,
    Country_Code,
    New_Rank,
    Mean_GDP,
    Temperature,
    Wind_Speed
    ) 
    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
    '''

# Create a function to insert data into the table
def insert_data(cursor, data, sql_insert_query):
    try:
        # Execute the query
        cursor.execute(sql_insert_query, data)
    except mariadb.Error as e:
        print(f"Error inserting data: {e}")

#### Loop through the dataframe rows and insert data


for index, row in df.iterrows():
    # print(tuple(row))
    insert_data(cur, tuple(row), sql_insert_query)

conn.commit()

# Execute a SELECT statement to fetch data from tbl_Airquality
cur.execute("SELECT * FROM tbl_Airquality")

# Print Result-set
for (Year,
     Country,
     Population,
     Populated_Area_km2, Air_Pollution_Average_ug_m3_NO2,
     Air_Pollution_Average_ug_m3_O3, Air_Pollution_Average_ug_m3_PM10, Air_Pollution_Average_ug_m3_PM2_5,
     Annual_AQI_NO2,
     Annual_AQI_O3,
     Annual_AQI_PM10,
     Annual_AQI_PM2_5,
     Premature_Deaths_NO2,
     Premature_Deaths_O3,
     Premature_Deaths_PM10,
     Premature_Deaths_PM2_5, Years_Of_Life_Lost_NO2,
     Years_Of_Life_Lost_O3,
     Years_Of_Life_Lost_PM10, Years_Of_Life_Lost_PM2_5,
     RenewableData,
     Average_Renewable_Data,
     coordinates,
     elevation,
     Country_Code,
     New_Rank,
     Mean_GDP,
     Temperature,
     Wind_Speed) in cur:
    print(f"Year: {Year}, Country: {Country}, Population: {Population}, "
          f"Populated Area (km^2): {Populated_Area_km2}, NO2 (ug/m3): {Air_Pollution_Average_ug_m3_NO2}, "
          f"O3 (ug/m3): {Air_Pollution_Average_ug_m3_O3}, PM10 (ug/m3): {Air_Pollution_Average_ug_m3_PM10}, "
          f"PM2.5 (ug/m3): {Air_Pollution_Average_ug_m3_PM2_5}, Annual AQI NO2: {Annual_AQI_NO2}, "
          f"Annual AQI O3: {Annual_AQI_O3}, Annual AQI PM10: {Annual_AQI_PM10}, "
          f"Annual AQI PM2.5: {Annual_AQI_PM2_5}, Premature Deaths NO2: {Premature_Deaths_NO2}, "
          f"Premature Deaths O3: {Premature_Deaths_O3}, Premature Deaths PM10: {Premature_Deaths_PM10}, "
          f"Premature Deaths PM2.5: {Premature_Deaths_PM2_5}, Years Of Life Lost NO2: {Years_Of_Life_Lost_NO2}, "
          f"Years Of Life Lost O3: {Years_Of_Life_Lost_O3}, Years Of Life Lost PM10: {Years_Of_Life_Lost_PM10}, "
          f"Years Of Life Lost PM2.5: {Years_Of_Life_Lost_PM2_5}, Renewable Data: {RenewableData}, "
          f"Average Renewable Data: {Average_Renewable_Data}",
          f"Coordinates: {coordinates}, Elevation: {elevation}, Country Code: {Country_Code}, New Rank: {New_Rank}, Mean GDP: {Mean_GDP}, Temperature: {Temperature}, Wind Speed: {Wind_Speed}")

# Check the number of columns
cur.execute(
    """
SELECT COUNT(*)
FROM INFORMATION_SCHEMA.COLUMNS
WHERE TABLE_SCHEMA = 'AIRQUALITY' AND TABLE_NAME = 'tbl_Airquality'
"""
)
column_count = cur.fetchone()[0]
print(f"Number of columns in 'tbl_Airquality': {column_count}")

# Check the Number of Rows
cur.execute("SELECT COUNT(*) FROM tbl_Airquality")
row_count = cur.fetchone()[0]
print(f"Number of rows in 'tbl_Airquality': {row_count}")


# Close Connection
conn.close()


### Prepare data



data_to_insert = [
    ("Belarus", "53.5318804783, 28.033566395", 170.0),
    ("Germany", "51.1063634863, 10.3814938434", 263.0),
    ("Denmark", "62.0740272545, -6.87055477617", 34.0),
    ("Portugal", "39.5871468639, -8.59083758901", 372.0),
    ("Andorra", "42.541326525, 1.56075624107", 1996.0),
    ("Albania", "41.1424820604, 20.053819197", 708.0),
    ("Poland", "52.1247718189, 19.3943221398", 173.0),
    ("Bosnia and Herzegovina", "44.1717340565, 17.7734080203", 500.0),
    ("United Kingdom", "49.2210642111, -2.12743081446", 162.0),
    ("Lithuania", "55.3270236412, 23.8882490157", 110.0),
    ("San Marino", "43.9364661946, 12.4422225588", 538.0),
    ("France", "42.1801069806, -2.75522118992", 375.0),
    ("Spain", "40.2272337014, -3.65263701695", 660.0),
    ("Hungary", "47.1632436866, 19.3967592674", 143.0),
    ("Croatia", "45.067726168, 16.3985488378", 331.0),
    ("Monaco", "43.7398461874, 7.39897432483", 375.0),
    ("Finland", "60.1864472514, 20.0555828218", 164.0),
    ("Czech Republic", "49.7324482055, 15.3141031033", 430.0),
    ("Liechtenstein", "47.1382064295, 9.54090028864", 1350.0),
    ("Italy", "42.7864339964, 12.0719070927", 538.0),
    ("Montenegro", "42.7858180546, 19.2385002864", 1086.0),
    ("Ukraine", "49.1600782909, 31.2553061432", 175.0),
    ("Sweden", "62.7770998648, 16.7544664339", 320.0),
    ("Greece", "39.0511811824, 22.9748189255", 498.0),
    ("Slovenia", "46.1161109843, 14.8080896248", 492.0),
    ("Luxembourg", "49.7668242758, 6.07176396727", 325.0),
    ("Belgium", "50.6374616353, 4.63996050767", 181.0),
    ("Estonia", "58.6710141504, 25.5357932075", 61.0),
    ("Kosovo", "42.5728833853, 20.8737431935", 800.0),
    ("Austria", "47.5856569205, 14.1306728873", 910.0),
    ("Netherlands", "52.0697916544, 5.2168142092", 30.0),
    ("Latvia", "56.8518194796, 24.9086224276", 87.0),
    ("Norway", "68.7942729218, 15.4095991585", 460.0),
    ("Romania", "45.8534640978, 24.9729418552", 414.0),
    ("Switzerland", "46.7989568698, 8.21535365745", 1350.0),
    ("Russia", "61.9668529189, 96.6275884228", 600.0),
    ("Moldova", "47.195101489, 28.4613971304", 139.0),
    ("Iceland", "64.9983158205, -18.5964459278", 557.0),
    ("Serbia", "44.2213759212, 20.790402471", 473.0),
    ("Bulgaria", "42.7685226144, 25.2175125117", 470.0),
    ("Slovakia", "48.7049353017, 19.4803829712", 458.0),
    ("Vatican City", "41.9033124692, 12.4534177294", 538.0),
    ("Ireland", "53.1782213727, -8.14504467255", 118.0),
    ("North Macedonia", "41.5950049267, 21.6838304101", 741.0)
]
